libcloud: Add MeanShift algorithm in order to locate maxima on Hough parameter space

+++ libcloud/2D/GaussianEstimator.cpp
@@ -14,7 +14,7 @@ GaussianEstimator::~GaussianEstimator ()
 }
 
 void
-GaussianEstimator::compute (const Matrix<UInt8>& input, Matrix<UInt8>& output) const
+GaussianEstimator::compute (const Matrix<UInt32>& input, Matrix<UInt32>& output) const
 {
   input.convolve (gaussian_filter, output);
 }
